src/main/java/zda/task/tasktracker/controller/LoginController.java
package zda.task.tasktracker.controller;

import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import javax.servlet.http.HttpServletRequest;

@Controller
public class LoginController {

    @RequestMapping("/main")
    public String toMainPage(Model model) {
        model.addAttribute("username", getUserName());
        return "main";
    }

    @RequestMapping("/")
    public String toLoginPage() {
        return "redirect:/login";
    }

    @RequestMapping("/logout")
    public String logout(HttpServletRequest request) {
        request.getSession().invalidate();

        return "redirect:/";
    }

    public String getUserName() {
        return SecurityContextHolder.getContext().getAuthentication().getName();
    }
}